Moved includes into common folder
